PNTN-220 event model split & refactor
