+++ Mimir/src/Model.php
@@ -40,4 +40,13 @@ abstract class Model
         $this->_config = $config;
         $this->_meta = $meta;
     }
+
+    /**
+     * Check if token allows administrative operations
+     * @return bool
+     */
+    public function checkAdminToken()
+    {
+        return $this->_meta->getAuthToken() === $this->_config->getValue('admin.god_token');
+    }
 }
