+++ Mimir/src/controllers/Seating.php
@@ -19,6 +19,8 @@ namespace Mimir;
 
 require_once __DIR__ . '/../Controller.php';
 require_once __DIR__ . '/../helpers/Seating.php';
+require_once __DIR__ . '/../models/Event.php';
+require_once __DIR__ . '/../models/EventRatingTable.php';
 
 class SeatingController extends Controller
 {
@@ -31,6 +33,7 @@ class SeatingController extends Controller
      * @param int $seed
      * @throws InvalidParametersException
      * @throws AuthFailedException
+     * @throws \Exception
      * @return bool
      */
     public function makeShuffledSeating($eventId, $groupsCount, $seed)
@@ -61,6 +64,7 @@ class SeatingController extends Controller
      * @param int $eventId
      * @throws InvalidParametersException
      * @throws AuthFailedException
+     * @throws \Exception
      * @return bool
      */
     public function makeSwissSeating($eventId)
@@ -116,6 +120,7 @@ class SeatingController extends Controller
      * @throws DatabaseException
      * @throws InvalidParametersException
      * @throws InvalidUserException
+     * @throws \Exception
      * @return bool
      */
     public function makeManualSeating($eventId, $tablesDescription, $randomize = false)
@@ -149,6 +154,7 @@ class SeatingController extends Controller
      * @throws DatabaseException
      * @throws InvalidParametersException
      * @throws InvalidUserException
+     * @throws \Exception
      * @return bool
      */
     public function makeIntervalSeating($eventId, $step)
@@ -162,7 +168,7 @@ class SeatingController extends Controller
             throw new InvalidParametersException('Event id#' . $eventId . ' not found in DB');
         }
 
-        $currentRatingTable = (new EventModel($this->_db, $this->_config, $this->_meta))
+        $currentRatingTable = (new EventRatingTableModel($this->_db, $this->_config, $this->_meta))
             ->getRatingTable($event[0], 'rating', 'desc');
 
         $seating = Seating::makeIntervalSeating($currentRatingTable, $step, true);
@@ -184,6 +190,7 @@ class SeatingController extends Controller
      * Update event "seating ready" status.
      * This should be done before games start, or admin panel will show some inadequate data.
      * @param $eventId
+     * @throws \Exception
      * @return bool flag if games are started immediately
      */
     protected function _updateEventStatus($eventId)
@@ -202,6 +209,7 @@ class SeatingController extends Controller
      * @param $eventId
      * @throws AuthFailedException
      * @throws InvalidParametersException
+     * @throws \Exception
      */
     protected function _checkIfAllowed($eventId)
     {
@@ -219,6 +227,7 @@ class SeatingController extends Controller
      * @param $eventId
      * @return array
      * @throws InvalidParametersException
+     * @throws \Exception
      */
     protected function _getData($eventId)
     {
@@ -270,6 +279,7 @@ class SeatingController extends Controller
      * @param $tablesDescription
      * @param $randomize
      * @throws InvalidParametersException
+     * @throws \Exception
      * @return array
      */
     protected function _makeManualSeating($eventId, $tablesDescription, $randomize)
@@ -288,7 +298,7 @@ class SeatingController extends Controller
             throw new InvalidParametersException('Event id#' . $eventId . ' not found in DB');
         }
 
-        $currentRatingTable = (new EventModel($this->_db, $this->_config, $this->_meta))
+        $currentRatingTable = (new EventRatingTableModel($this->_db, $this->_config, $this->_meta))
             ->getRatingTable($event[0], 'rating', 'desc');
 
         $participatingPlayers = [];
