+++ TODO.txt
@@ -1,0 +1,33 @@
+- Предопределенная рассадка на несколько туров:
+    - Делаем отдельную сущность (EventPrescript?) (готово)
+    - В сущности храним: 1) сам скрипт, 2) текущий статус скрипта в турнире (какая игра сыграна) (готово)
+    - Сопоставление игрока и локального идентификатора в турнире делается через примитив EventRegistration (готово)
+    - Делаем страницу редактирования скрипта. На странице можно ввести сам скрипт в текстовом виде и задать текущий статус (номер сыгранной игры).
+    - Делаем в разделе управления пользователями рядом с кнопками перевыдачи пинов поля для задания локального айди
+    - Делаем в разделе управления пользователями галочки "Заменен", на случай временной замены. Эта функциональность должна быть доступна и в нескриптовом режиме!
+    - * Следствие вышеуказанных условий: при необходимости постоянной замены игрока - нужно убрать его из евента и добавить другого с тем же local_id.
+    - В управлении турнирами для скриптового режима скрываем все рассадки, показываем единственную кнопку "следующий тур". Также возможно следует показать там таблицу игроков для контроля админом состава столов.
+    - Если все туры сыграны - не показываем кнопку и не даем ничего сделать кроме скрытия/показа рейтинговой таблицы.
+    
+
+
+
+p1
+- GUI: добавить пункты статы - 
+                Ничьи
+                  - из них в темпае
+                  - из них в нотене
+- Статистика:
+        - добавить стоимость рук в которые игрок накидывал по рону
+
+p2
+- Более универсальная авторизация. Вероятно - через вк. Иные более универсальные средства слишком дороги.
+- Проверять, что один и тот же игрок не играет за двумя столами?
+- Подумать над отдельным экраном выбора риичи-ставок
+- Просмотр промежуточных итогов - сомнительно, но возможно имеет смысл
+- Вместо якухай N сделать хаку, хацу, чун и ветра; скорее всего настройкой (не везде это нужно, на турнирах например лучше быстро вбить якухай 3).
+- Фичреквест - если я выбираю яку "риичи", ставка ставится автоматически (т.е. наоборот относительно того что сейчас).
+
+
+? Риншан по взятия в кан с кого-то - как это обрабатывать???
+? Как обработать сетевые ошибки? Как вообще строить обработку ошибок?
