--- Tyr/src/app/TODO.txt
@@ -1,21 +1,0 @@
-- Переключатель языков через значение в localstorage
-
-p1
-- GUI: добавить пункты статы - 
-                Ничьи
-                  - из них в темпае
-                  - из них в нотене
-- Статистика:
-        - добавить стоимость рук в которые игрок накидывал по рону
-
-p2
-- Более универсальная авторизация. Вероятно - через вк. Иные более универсальные средства слишком дороги.
-- Проверять, что один и тот же игрок не играет за двумя столами?
-- Подумать над отдельным экраном выбора риичи-ставок
-- Просмотр промежуточных итогов - сомнительно, но возможно имеет смысл
-- Вместо якухай N сделать хаку, хацу, чун и ветра; скорее всего настройкой (не везде это нужно, на турнирах например лучше быстро вбить якухай 3).
-- Фичреквест - если я выбираю яку "риичи", ставка ставится автоматически (т.е. наоборот относительно того что сейчас).
-
-
-? Риншан по взятия в кан с кого-то - как это обрабатывать???
-? Как обработать сетевые ошибки? Как вообще строить обработку ошибок?
